beginning of tests (with unittest) for integration time
