+++ gui/views/microRamanView.py
@@ -707,9 +707,9 @@ class MicroRamanView(QWidget, Ui_microRamanView):  # type: QWidget
                 self.pb_saveImage.setEnabled(True)
                 self.cmb_wave.setEnabled(False)
                 self.disable_all_buttons()
-                self.set_integration_time()
                 self.create_plot_rgb()
                 self.create_plot_spectrum()
+                self.set_exposure_time()
                 self.spectrum_pixel_acquisition()
                 self.create_matrix_raw_data()
                 self.create_matrix_rgb()
@@ -723,8 +723,8 @@ class MicroRamanView(QWidget, Ui_microRamanView):  # type: QWidget
             if self.countSpectrum <= (self.width*self.height):
                 if self.countHeight != 0 or self.countWidth != 0:
                     self.spectrum_pixel_acquisition()
-                self.matrix_raw_data_replace()
 
+                self.matrix_raw_data_replace()
                 self.matrixRGB_replace()
                 self.update_rgb_plot()
